Summaries

Summaries

Here, we summarize all types definitions, operations and functions supported by Liga package.

Summary of types

Liga.kbasis — Type.

c(x::Vector{Bool}, a::Number)

Creates a scaled basis element of geometric algebra space from x and a.

julia> kbasis(e13, 2.3)
2.3e13
julia> kbasis(e23)
1.0e23

source

Liga.kmultvec — Type.

kmultvec(X::Vector{kbasis})

Creates a multivector in the Euclidean geometric algebra (Gₚ) composed by the sum of the coordinates of X.

julia> kmultvec([kbasis(e1, 2.0), kbasis(e12, 3.3), kbasis(e123)])
2.0e1 + 3.3e12 + 1.0e123

source

Liga.kblade — Type.

kblade(X::Vector{kmultvec})

Creates a blade in the Euclidean geometric algebra (Gp) composed by the outer product of the elements of X.

The blade is created only if the multivectors that compound X are both 1-vectors and X is a L.i set

julia> a = kmultvec([kbasis(e1,2.0),kbasis(e3,4.0),kbasis(e3,4.0)])
2.0e1 + 4.0e3 + 4.0e3
julia> b = kmultvec([kbasis(e1, 1.0), kbasis(e2, 1.0), kbasis(e3,2.0)])
1.0e1 + 1.0e2 + 2.0e3
julia> A = kblade([a,b])
(2.0e1 + 4.0e3 + 4.0e3)∧(1.0e1 + 1.0e2 + 2.0e3)

source

Liga.cbasis — Type.

cbasis(a::Vector{Bool},b::Bool,c::Bool,d::Number)

Creates a scaled basis element of geometric algebra space (Gp+1,1) from a, b, c and d using e∞ and e∘ as basis elements instead of the regular basis.

source

Liga.cmultvec — Type.

cmultvec(X::Vector{cbasis})

Creates a multivector in the geometric algebra (Gp+1,1) composed by the sum of the coordinates of X.

source

Liga.cblade — Type.

cblade(X::Vector{cmultvec})

Creates a blade in the geometric algebra $ \mathbb{G}_{k+1,1} $ composed by the outer product of the elements of X.

source

Summary of functions

Liga.geoprod — Function.

geoprod(a::kbasis, b::kbasis)
geoprod(X::kmultvec, Y::kmultvec)
geoprod(A::kblade, B::kblade)
geoprod(a::pbasis, b::pbasis)
geoprod(X::pmultvec, Y::pmultvec)
geoprod(A::pblade, B::pblade)
geoprod(a::cbasis, b::cbasis)
geoprod(X::cmultvec, Y::cmultvec)
geoprod(A::cblade, B::cblade)
geoprod(A::Array{kbasis,2},B::Array{kbasis,1})
geoprod(A::Array{kbasis,2},B::Array{kbasis,2})
geoprod(A::Array{kmultvec,2},B::Array{kmultvec,1})
geoprod(A::Array{kmultvec,2},B::Array{kmultvec,2})

Receives as input parameters elements of the same type and returns the geometric product between them. This function is extended to array like a traditional product. It can be used the operator "∘ (circ)" instead of call the function.

source

Liga.inner — Function.

inner(a::kbasis, b::kbasis)
inner(X::kmultvec, Y::kmultvec)
inner(A::kblade, B::kblade)
inner(a::pbasis, b::pbasis)
inner(X::pmultvec, Y::pmultvec)
inner(A::pblade, B::pblade)
inner(a::cbasis, b::cbasis)
inner(X::cmultvec, Y::cmultvec)
inner(A::cblade, B::cblade)

Receives as input parameters elements of the same type and returns the inner product between them.

It can be used the operator "⋅ (cdot)" instead of call the function.

source

Liga.outer — Function.

outer(a::kbasis, b::kbasis)
outer(X::kmultvec, Y::kmultvec)
outer(A::kblade, B::kblade)
outer(a::pbasis, b::pbasis)
outer(X::pmultvec, Y::pmultvec)
outer(A::pblade, B::pblade)
outer(a::cbasis, b::cbasis)
outer(X::cmultvec, Y::cmultvec)
outer(A::cblade, B::cblade)

Receives as input parameters elements of the same type and returns the outer product between them.

It can be used the operator " ^ " instead of call the function.

source

Liga.scalar — Function.

scalar(a::kbasis, b::kbasis)
scalar(X::kmultvec, Y::kmultvec)
scalar(A::kblade, B::kblade)
scalar(a::pbasis, b::pbasis)
scalar(X::pmultvec, Y::pmultvec)
scalar(A::pblade, B::pblade)
scalar(a::cbasis, b::cbasis)
scalar(X::cmultvec, Y::cmultvec)
scalar(A::cblade, B::cblade)

Receives as input parameters elements of the same type and returns the scalar product between them.

It can be used the operator " * " instead of call the function.

source

Base.transpose — Method.

transpose(A::Array{kbasis,1})
transpose(A::Array{kbasis,2})
transpose(A::Array{kmultvec,1})
transpose(A::Array{kmultvec,2})

Returns the transpose of an Array defined by kbasis or kmultvec elements.

source

Liga.dual — Function.

dual(a::kbasis)
dual(X::kmultvec)
dual(A::kblade)
dual(X::pmultvec)
dual(a::pbasis)
dual(A::pblade)

Returns the dual element of the basis blade a, the multivetor X or the blade A.

source

Liga.grade — Function.

grade(a::kbasis)
grade(a::pbasis)
grade(a::cbasis)

Returns the grade of the basis element x

julia> grade(kbasis(e13, 1.5))
2

source

Liga.inverse — Function.

inverse(a::kmultvec)

Returns the inverse of a multi vector input. Be careful with this function because in current version it isn't optimized!

source

inverse(A::kblade)
inverse(A::pblade)
inverse(A::cblade)

Returns the inverse of A if A is a non-null-blade or the pseudoinverse A if A is a null-blade.

source

Liga.mvsum — Function.

mvsum(a::kbasis, b::kbasis)
mvsum(X::kMulvec, Y::kmultvec)
mvsum(a::pbasis, b::pbasis)
mvsum(X::pMulvec, Y::pmultvec)
mvsum(a::cbasis, b::cbasis)
mvsum(X::cMulvec, Y::cmultvec)

Receives as input parameters elements of the same type and returns the sum of them.

It can be used the operators "+" or "-" instead of call the function.

source

Liga.mvreverse — Function.

mvreverse(a::kbasis)
mvreverse(X::kmultvec)
mvreverse(A::kblade)
mvreverse(a::pbasis)
mvreverse(A::pblade)
mvreverse(a::cbasis)
mvreverse(A::cblade)

Returns the reverse of the element.

source

Liga.magnitude — Function.

magnitude(a::kbasis)
magnitude(a::kmultvec)
magnitude(A::kblade)
magnitude(a::pbasis)
magnitude(a::pmultvec)
magnitude(A::pblade)
magnitude(a::cbasis)
magnitude(A::cblade)

Returns the magnitude of the input element.

source

Liga.projection — Function.

projection(A::kblade, N::kblade)
projection(A::pblade, N::pblade)

Returns the projection of the blade A onto the blade N.

source

Liga.rejection — Function.

rejection(A::kblade, N::kblade)
rejection(A::pblade, N::pblade)

Returns the rejection of the blade A from the blade N.

source

Liga.S — Function.

S(x::Vector{Float64})
S(X::kmultvec)

The stereographic embedding of Euclidean space.

Takes a vector of Euclidean space in its stereographic embedding.

source

Liga.Hm — Function.

Hm(x::Vector{Float64})

Auxiliary function.

source

Liga.H — Function.

H(X::Vector{Float64})
H(X::kmultvec)

Homegenization. It takes a vector in $mathbb{R}^n$ into its embedding in the affine space.

source

Liga.iH — Function.

iH(X::kmultvec)

Inversion for homonezation function H.

source

Liga.pconformal — Function.

pconformal(x::Vector{Float64})

Conformal embedding of a Euclidean vector into a vector of Gp+1,1 with new basis elements being e+ and e-.

source

Liga.ipconformal — Function.

ipconformal(X::pmultvec)

The inverse of "pconformal" function.

source

Liga.conformal — Function.

conformal(x::Vector{Float64})

Conformal embedding of a Euclidean vector into a vector of Gp+1,1 with new basis elements being e∞ and e∘.

source

Liga.iconformal — Function.

iconformal(X::cmultvec)

The inverse of "conformal" function.

source